ENH: Restructure custom ops (#49)
* ENH: Restructure custom ops
