+++ BUILD
@@ -6,10 +6,9 @@ sh_binary(
         "MANIFEST.in",
         "setup.py",
         "tensorflow_addons/__init__.py",
-        "//tensorflow_addons/image:images_ops_py",
+        "//tensorflow_addons/custom_ops:custom_ops_py",
         "//tensorflow_addons/layers:layers_py",
         "//tensorflow_addons/losses:losses_py",
         "//tensorflow_addons/optimizers:optimizers_py",
-        "//tensorflow_addons/text:text_py",
     ],
 )
