+++ tensorflow_addons/custom_ops/README.md
@@ -1,0 +1,22 @@
+# Addons - Custom Ops
+
+## Contents
+| Layer  | Description                             |
+|:----------------------- |:-----------------------------|
+| Image | Ops for image manipulation   |
+| Text |  Ops for text processing  |
+
+
+
+## Contribution Guidelines
+#### Standard API
+In order to conform with the current API standard, all custom ops
+must:
+ * Must be impossible to implement in one of the other API
+ standards (Layers, Losses, etc.)
+
+#### Testing Requirements
+ * Simple unittests that demonstrate the custom op is behaving as
+    expected.
+ * When applicable, run all unittests with TensorFlow's
+  `@run_all_in_graph_and_eager_modes` decorator.
