+++ tensorflow_addons/custom_ops/image/cc/ops/image_ops.cc
