+++ tensorflow_addons/custom_ops/text/__init__.py
@@ -12,12 +12,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
-"""Additional layers that conform to Keras API"""
-
+"""Text-processing ops"""
 from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-from tensorflow_addons.layers.python.maxout import Maxout
-from tensorflow_addons.layers.python.poincare import PoincareNormalize
-from tensorflow_addons.layers.python.wrappers import WeightNormalization
+# Skip Gram Sampling
+from tensorflow_addons.custom_ops.text.python.skip_gram_ops import skip_gram_sample
+from tensorflow_addons.custom_ops.text.python.skip_gram_ops import skip_gram_sample_with_text_vocab
