--- tensorflow_addons/image/README.md
@@ -1,4 +1,0 @@
-# Addons - Image
-
-
-## Standard API
