--- tensorflow_addons/text/README.md
@@ -1,3 +1,0 @@
-# Addons Text
-
-## Standard API
